requirements.txt
flask
requests
python-whois
dnspython